Solved linter errors
